Moves alarm notifications from installation to configuration.
