+++ osgp-adapter-domain-smartmetering/src/main/java/com/alliander/osgp/adapter/domain/smartmetering/application/mapping/ConfigurationMapper.java
@@ -20,9 +20,12 @@ public class ConfigurationMapper extends ConfigurableMapper {
 
         // domain value object -> dto value object
         mapperFactory
-                .classMap(com.alliander.osgp.domain.core.valueobjects.smartmetering.SpecialDaysRequest.class,
-                        com.alliander.osgp.dto.valueobjects.smartmetering.SpecialDaysRequest.class).byDefault()
+        .classMap(com.alliander.osgp.domain.core.valueobjects.smartmetering.SpecialDaysRequest.class,
+                com.alliander.osgp.dto.valueobjects.smartmetering.SpecialDaysRequest.class).byDefault()
                 .register();
 
+        mapperFactory
+        .classMap(com.alliander.osgp.domain.core.valueobjects.smartmetering.AlarmSwitches.class,
+                com.alliander.osgp.dto.valueobjects.smartmetering.AlarmSwitches.class).byDefault().register();
     }
 }
