+++ osgp-adapter-domain-smartmetering/src/main/java/com/alliander/osgp/adapter/domain/smartmetering/application/services/ConfigurationService.java
@@ -19,10 +19,14 @@ import org.springframework.transaction.annotation.Transactional;
 import com.alliander.osgp.adapter.domain.smartmetering.application.mapping.ConfigurationMapper;
 import com.alliander.osgp.adapter.domain.smartmetering.infra.jms.core.OsgpCoreRequestMessageSender;
 import com.alliander.osgp.adapter.domain.smartmetering.infra.jms.ws.WebServiceResponseMessageSender;
+import com.alliander.osgp.domain.core.entities.SmartMeteringDevice;
+import com.alliander.osgp.domain.core.repositories.SmartMeteringDeviceRepository;
 import com.alliander.osgp.domain.core.validation.Identification;
+import com.alliander.osgp.domain.core.valueobjects.smartmetering.AlarmSwitches;
 import com.alliander.osgp.dto.valueobjects.smartmetering.SpecialDaysRequest;
 import com.alliander.osgp.shared.exceptionhandling.ComponentType;
 import com.alliander.osgp.shared.exceptionhandling.FunctionalException;
+import com.alliander.osgp.shared.exceptionhandling.FunctionalExceptionType;
 import com.alliander.osgp.shared.exceptionhandling.OsgpException;
 import com.alliander.osgp.shared.exceptionhandling.TechnicalException;
 import com.alliander.osgp.shared.infra.jms.RequestMessage;
@@ -44,6 +48,9 @@ public class ConfigurationService {
 
     @Autowired
     private WebServiceResponseMessageSender webServiceResponseMessageSender;
+
+    @Autowired
+    private SmartMeteringDeviceRepository smartMeteringDeviceRepository;
 
     public ConfigurationService() {
         // Parameterless constructor required for transactions...
@@ -99,6 +106,56 @@ public class ConfigurationService {
                     "Unexpected exception while retrieving response message", e);
 
         }
+        this.webServiceResponseMessageSender.send(new ResponseMessage(correlationUid, organisationIdentification,
+                deviceIdentification, result, osgpException, null), messageType);
+    }
+
+    public void setAlarmNotifications(@Identification final String organisationIdentification,
+            @Identification final String deviceIdentification, final String correlationUid,
+            final AlarmSwitches alarmSwitches, final String messageType) throws FunctionalException {
+
+        LOGGER.info("setAlarmNotifications for organisationIdentification: {} for deviceIdentification: {}",
+                organisationIdentification, deviceIdentification);
+
+        // TODO: bypassing authorization, this should be fixed.
+        // Organisation organisation =
+        // this.findOrganisation(organisationIdentification);
+        // final Device device = this.findActiveDevice(deviceIdentification);
+
+        final SmartMeteringDevice device = this.smartMeteringDeviceRepository
+                .findByDeviceIdentification(deviceIdentification);
+        if (device == null) {
+            throw new FunctionalException(FunctionalExceptionType.UNKNOWN_DEVICE, ComponentType.DOMAIN_SMART_METERING);
+        }
+
+        final com.alliander.osgp.dto.valueobjects.smartmetering.AlarmSwitches alarmSwitchesDto = this.configurationMapper
+                .map(alarmSwitches, com.alliander.osgp.dto.valueobjects.smartmetering.AlarmSwitches.class);
+
+        this.osgpCoreRequestMessageSender.send(new RequestMessage(correlationUid, organisationIdentification,
+                deviceIdentification, alarmSwitchesDto), messageType);
+    }
+
+    public void handleSetAlarmNotificationsResponse(final String deviceIdentification,
+            final String organisationIdentification, final String correlationUid, final String messageType,
+            final ResponseMessageResultType deviceResult, final OsgpException exception) {
+
+        LOGGER.info("handleSetAlarmNotificationsResponse for MessageType: {}", messageType);
+
+        ResponseMessageResultType result = ResponseMessageResultType.OK;
+        OsgpException osgpException = exception;
+
+        try {
+            if (deviceResult == ResponseMessageResultType.NOT_OK || osgpException != null) {
+                LOGGER.error("Set Alarm Notifications Response not ok.", osgpException);
+                throw osgpException;
+            }
+        } catch (final Exception e) {
+            LOGGER.error("Unexpected Exception", e);
+            result = ResponseMessageResultType.NOT_OK;
+            osgpException = new TechnicalException(ComponentType.UNKNOWN,
+                    "Unexpected exception while retrieving response message", e);
+        }
+
         this.webServiceResponseMessageSender.send(new ResponseMessage(correlationUid, organisationIdentification,
                 deviceIdentification, result, osgpException, null), messageType);
     }
