+++ osgp-adapter-ws-smartmetering/src/main/java/com/alliander/osgp/adapter/ws/smartmetering/application/mapping/InstallationMapper.java
@@ -13,7 +13,6 @@ import ma.glasnost.orika.impl.ConfigurableMapper;
 import org.springframework.stereotype.Component;
 
 import com.alliander.osgp.adapter.ws.schema.smartmetering.installation.Device;
-import com.alliander.osgp.domain.core.valueobjects.smartmetering.AlarmType;
 import com.alliander.osgp.domain.core.valueobjects.smartmetering.SmartMeteringDevice;
 
 @Component(value = "installationMapper")
@@ -21,8 +20,5 @@ public class InstallationMapper extends ConfigurableMapper {
     @Override
     public void configure(final MapperFactory mapperFactory) {
         mapperFactory.classMap(Device.class, SmartMeteringDevice.class).byDefault().register();
-        mapperFactory
-                .classMap(com.alliander.osgp.adapter.ws.schema.smartmetering.installation.AlarmType.class,
-                        AlarmType.class).byDefault().register();
     }
 }
