+++ osgp-adapter-ws-smartmetering/src/main/java/com/alliander/osgp/adapter/ws/smartmetering/application/services/InstallationService.java
@@ -18,7 +18,6 @@ import com.alliander.osgp.adapter.ws.smartmetering.infra.jms.SmartMeteringReques
 import com.alliander.osgp.adapter.ws.smartmetering.infra.jms.SmartMeteringRequestMessageType;
 import com.alliander.osgp.domain.core.services.CorrelationIdProviderService;
 import com.alliander.osgp.domain.core.validation.Identification;
-import com.alliander.osgp.domain.core.valueobjects.smartmetering.AlarmSwitches;
 import com.alliander.osgp.domain.core.valueobjects.smartmetering.SmartMeteringDevice;
 import com.alliander.osgp.shared.exceptionhandling.FunctionalException;
 
@@ -80,45 +79,5 @@ public class InstallationService {
     public String addDevice(final String organisationIdentification, final SmartMeteringDevice device)
             throws FunctionalException {
         return this.enqueueAddSmartMeterRequest(organisationIdentification, device.getDeviceIdentification(), device);
-    }
-
-    public String enqueueSetAlarmNotificationsRequest(@Identification final String organisationIdentification,
-            @Identification final String deviceIdentification, final AlarmSwitches alarmSwitches)
-            throws FunctionalException {
-
-        /*
-         * TODO check if something needs to be done for authorization logic
-         * 
-         * In enqueueAddSmartMeterRequest some comments are made about the
-         * authorization being bypassed for now, that may be applicable in this
-         * method also.
-         */
-
-        LOGGER.debug("enqueueSetAlarmNotificationsRequest called with organisation {} and device {}",
-                organisationIdentification, deviceIdentification);
-
-        final String correlationUid = this.correlationIdProviderService.getCorrelationId(organisationIdentification,
-                deviceIdentification);
-
-        final SmartMeteringRequestMessage message = new SmartMeteringRequestMessage(
-                SmartMeteringRequestMessageType.SET_ALARM_NOTIFICATIONS, correlationUid, organisationIdentification,
-                deviceIdentification, alarmSwitches);
-
-        this.smartMeteringRequestMessageSender.send(message);
-
-        return correlationUid;
-    }
-
-    /**
-     * @param organisationIdentification
-     * @param deviceIdentification
-     * @param enableAlarms
-     * @param disableAlarms
-     * @throws FunctionalException
-     */
-    public String setAlarmNotifications(final String organisationIdentification, final String deviceIdentification,
-            final AlarmSwitches alarmSwitches) throws FunctionalException {
-        return this
-                .enqueueSetAlarmNotificationsRequest(organisationIdentification, deviceIdentification, alarmSwitches);
     }
 }
