+++ osgp-adapter-ws-smartmetering/src/main/webapp/WEB-INF/wsdl/smartmetering/SmartMeteringInstallation.wsdl
@@ -33,20 +33,6 @@
         </wsdl:part>
     </wsdl:message>
 
-    <wsdl:message name="SetAlarmNotificationsHeader">
-        <wsdl:part element="common:OrganisationIdentification" name="OrganisationIdentification" />
-        <wsdl:part element="common:UserName" name="UserName" />
-		<wsdl:part element="common:ApplicationName" name="ApplicationName" />
-    </wsdl:message>
-    <wsdl:message name="SetAlarmNotificationsResponse">
-        <wsdl:part element="sminst:SetAlarmNotificationsResponse" name="SetAlarmNotificationsResponse">
-        </wsdl:part>
-    </wsdl:message>
-    <wsdl:message name="SetAlarmNotificationsRequest">
-        <wsdl:part element="sminst:SetAlarmNotificationsRequest" name="SetAlarmNotificationsRequest">
-        </wsdl:part>
-    </wsdl:message>
-
     <wsdl:portType name="SmartmeterInstallationPort">
    
         <wsdl:operation name="AddDevice">
@@ -55,13 +41,6 @@
             <wsdl:output message="tns:AddDeviceResponse" name="AddDeviceResponse">
             </wsdl:output>
         </wsdl:operation>
-
-		<wsdl:operation name="SetAlarmNotifications">
-            <wsdl:input message="tns:SetAlarmNotificationsRequest" name="SetAlarmNotificationsRequest">
-            </wsdl:input>
-            <wsdl:output message="tns:SetAlarmNotificationsResponse" name="SetAlarmNotificationsResponse">
-            </wsdl:output>
-		</wsdl:operation>
 
     </wsdl:portType>
     
@@ -77,19 +56,6 @@
                 <soap:body use="literal" />
             </wsdl:input>
             <wsdl:output name="AddDeviceResponse">
-                <soap:body use="literal" />
-            </wsdl:output>
-        </wsdl:operation>
-
-        <wsdl:operation name="SetAlarmNotifications">
-            <soap:operation soapAction="" />
-            <wsdl:input name="SetAlarmNotificationsRequest">
-                <soap:header message="tns:SetAlarmNotificationsHeader" part="OrganisationIdentification" use="literal" />
-                <soap:header message="tns:SetAlarmNotificationsHeader" part="UserName" use="literal" />
-				<soap:header message="tns:SetAlarmNotificationsHeader" part="ApplicationName" use="literal" />
-                <soap:body use="literal" />
-            </wsdl:input>
-            <wsdl:output name="SetAlarmNotificationsResponse">
                 <soap:body use="literal" />
             </wsdl:output>
         </wsdl:operation>
